Hip IK: suppress the IK mechanism around the singularity.
When ends of an IK chain are too close it becomes very unstable.
In this case just using limit distance produces a smaller error.

In real use this case should be rare, but it is quite likely
when enabled in the default T or A pose and looks very bad.