Spline Tentacle: support very approximate IK to FK snapping.
For this rig IK to FK snapping is not only ambiguous and hard,
but fundamentally unsolvable unless the FK shape itself was
obtained by FK to IK snapping.

Therefore, this operator simply snaps control points to equally
spaced locations along the FK curve to be used as a rough draft
for further manual tuning. It also tries to roughly approximate
base and tip twist. Except for the tip control in the Direct Tip
Control mode, scale is not supported at all.

Since this operation is only useful as a start for manual tweaking,
baking keyframes is not supported.